Fix AQI Calculation
Currently calculation breaks when the data point is less than or equal to the LOW value of the current data_range. It should be less than or equal to the high value of the current data_range, in order to select the correct data_range. Its very wrong at low concentrations.

A PM 2.5 concentration of 0.1 gives an AQI of 26+ when it should be 1, according to the EPA calculator. This change fixes it.

Fixes #4